test/ar/edu/unq/po/tp4/ProductoPrimeraNecesidadTest.java
package ar.edu.unq.po.tp4;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class ProductoPrimeraNecesidadTest {

	private ProductoPrimeraNecesidad leche;
	private ProductoPrimeraNecesidad sal;

	@BeforeEach
	public void setUp() {
		leche = new ProductoPrimeraNecesidad("Leche", 8d, false);
		sal = new ProductoPrimeraNecesidad("Sal", 6d, false, 20);
	}

	@Test
	public void testCalcularPrecio() {
		assertEquals(new Double(7.2), leche.getPrecio());
		assertEquals(new Double(4.8), sal.getPrecio());
	}
}